+++ src/presentation/handlers/auth.rs
@@ -189,6 +189,7 @@ pub async fn auth_change_password(
                     version: None,
                     action_type: Some("ADMIN"),
                     diff: None,
+                    stream: None,
                 },
             )
             .await?;
@@ -226,6 +227,7 @@ pub async fn auth_register(
             version: None,
             action_type: Some("ADMIN"),
             diff: None,
+            stream: None,
         },
     )
     .await?;
@@ -292,6 +294,7 @@ pub async fn create_token(
             version: None,
             action_type: Some("ADMIN"),
             diff: None,
+            stream: None,
         },
     )
     .await?;
@@ -320,6 +323,7 @@ pub async fn revoke_token(
             version: None,
             action_type: Some("ADMIN"),
             diff: None,
+            stream: None,
         },
     )
     .await?;